pyproject.toml
[project]
name = "remote-game-console"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "pyaudio>=0.2.14",
    "pygame>=2.6.1",
    "pyserial>=3.5",
]

[project.scripts]
remote-game-console = "remote_game_console.main:main"

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.14.0",
    "types-pyserial>=3.5.0.20251001",
]

[tool.uv]
package = true

[tool.mypy]
strict = true
